score cutoffs

+++ src/main/python/importer/scoreCalls.py
@@ -52,7 +52,7 @@ def loadCalledFile(fname,type):
 def getCountSmall(Li,scoreCutoff):
 	count = 0
 	for vs in Li:
-		if vs["type"].find("small")!=-1 and int(vs["score"]) >= scoreCutoff:
+		if vs["type"].find("small")!=-1 and float(vs["score"]) >= scoreCutoff:
 			count+=1
 	return count
 		
@@ -61,7 +61,7 @@ def getCountSmall(Li,scoreCutoff):
 def getCountEdge(Li,scoreCutoff):
 	count = 0
 	for vs in Li:
-		if vs["type"].find("other")!=-1 and int(vs["score"]) >= scoreCutoff:
+		if vs["type"].find("other")!=-1 and float(vs["score"]) >= scoreCutoff:
 			count+=1
 	return count
